requirements.txt
aiofiles==0.7.0
appdirs==1.4.4
argcomplete==1.12.3
asgiref==3.3.4
attrs==21.2.0
bcrypt==3.2.0
certifi==2021.5.30
cffi==1.15.0
chardet==4.0.0
click==8.0.1
colorama==0.4.4
dotty-dict==1.3.0
ecdsa==0.17.0
fastapi==0.65.2
h11==0.12.0
halo==0.0.31
hjson==3.0.2
idna==2.10
Jinja2==2.11.3
jsonschema==3.2.0
log-symbols==0.0.14
MarkupSafe==2.0.1
milc==1.3.0
passlib==1.7.4
Pillow==8.3.1
psycopg2==2.9.1
pyasn1==0.4.8
pycparser==2.21
pydantic==1.8.2
Pygments==2.9.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2021.1
requests==2.25.1
rsa==4.8
setuptools-scm==6.0.1
six==1.16.0
spinners==0.0.24
starlette==0.14.2
termcolor==1.1.0
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
